Move de_bruijn to proper file.
